add parser support for oneof and map fields.
I'm mostly interested in the parser and not in the C extensions generated so I've improved the parser while keeping its default behaviour intact.
I'm not a C guy so nothing I can help with adding the complete support for the new fields. I do hope it might help folks down the line add the full support.

Commit includes:
* add support for oneof field in parser.
* add support for map fields in parser.
* add capability to disable specific tokens in parser.
* disable the new oneof and map field by default in the parser to make sure the parser works with the C extensions out-of-the-box.
* add informative exceptions for disabled parser tokens.
* simplify the `tokenize` code by creating parsable tokens mapping from name to class.
* change a few function-level variable names to improve clarity