Pin jinja2 version to 2.11.3
Jinja2 v3.0 bumps its MarkupSafe dependency to a release candidate. This breaks resolution of the package on ARM-based machines. Since the changes in jinja 3.0 don't seem to affect the functionality of this project, we propose pinning the jinja version to 2.11.3 until its MarkupSafe dependency is bumped to a major version, instead of a release candidate.